# plot geometry for dispersive materials without initializing structure object