--- a/python/simulation.py
+++ b/python/simulation.py
@@ -4068,7 +4068,7 @@ class Simulation(object):
     def plot2D(self, ax=None, output_plane=None, fields=None, labels=False,
                eps_parameters=None, boundary_parameters=None,
                source_parameters=None, monitor_parameters=None,
-               field_parameters=None, frequency=None,
+               field_parameters=None, resolution=None,
                plot_eps_flag=True, plot_sources_flag=True,
                plot_monitors_flag=True, plot_boundaries_flag=True,
                **kwargs):
@@ -4089,13 +4089,18 @@ class Simulation(object):
         plt.show()
         plt.savefig('sim_domain.png')
         ```
+        If you just want to quickly visualize the simulation domain without the fields (i.e., when
+        debugging your simulation), there is no need to invoke the `run` function prior to calling
+        `plot2D`. Just define the `Simulation` object followed by any DFT monitors and then
+        invoke `plot2D`.
 
-        Note: When running a [parallel simulation](Parallel_Meep.md), the `plot2D` function expects to be called
-        on all processes, but only generates a plot on the master process.
+        Note: When running a [parallel simulation](Parallel_Meep.md), the `plot2D` function expects
+        to be called on all processes, but only generates a plot on the master process.
 
         Note: The geometry function is evaluated at infinite frequency (i.e., the instantaneous response).
-        Dispersive materials are ignored. To visualize dispersive materials, you can use the `get_array` or
-        `output_epsilon` functions for which a `frequency` parameter can be specified.
+        Dispersive materials containing complex, wavelength-dependent $\\varepsilon$ are ignored. To
+        visualize dispersive materials, you can use the `get_array` or `output_epsilon` functions and
+        specify a `frequency` parameter for evaluating the complex $\\varepsilon$.
 
         **Parameters:**
 
@@ -4117,6 +4122,8 @@ class Simulation(object):
             - `alpha=1.0`: transparency of geometry
             - `contour=False`: if `True`, plot a contour of the geometry rather than its image
             - `contour_linewidth=1`: line width of the contour lines if `contour=True`
+            - `resolution=None`: the resolution to sample the $\\varepsilon$ grid. Defaults to the
+               `resolution` of the `Simulation` object.
         * `boundary_parameters`: a `dict` of optional plotting parameters that override
           the default parameters for the boundary layers.
             - `alpha=1.0`: transparency of boundary layers
@@ -4164,6 +4171,7 @@ class Simulation(object):
                           source_parameters=source_parameters,
                           monitor_parameters=monitor_parameters,
                           field_parameters=field_parameters,
+                          resolution=resolution,
                           plot_eps_flag=plot_eps_flag,
                           plot_sources_flag=plot_sources_flag,
                           plot_monitors_flag=plot_monitors_flag,