--- a/python/visualization.py
+++ b/python/visualization.py
@@ -327,7 +327,7 @@ def plot_volume(sim, ax, volume, output_plane=None, plotting_parameters=None, la
             return ax
     return ax
 
-def plot_eps(sim, ax, output_plane=None, eps_parameters=None, resolution=None):
+def plot_eps(sim, ax, output_plane=None, eps_parameters=None, resolution=None, frequency=0):
     # consolidate plotting parameters
     eps_parameters = default_eps_parameters if eps_parameters is None else dict(default_eps_parameters, **eps_parameters)
 
@@ -376,7 +376,7 @@ def plot_eps(sim, ax, output_plane=None, eps_parameters=None, resolution=None):
     else:
         raise ValueError("A 2D plane has not been specified...")
 
-    eps_data = np.rot90(sim.get_epsilon_grid(xtics, ytics, ztics))
+    eps_data = np.rot90(np.real(sim.get_epsilon_grid(xtics, ytics, ztics, frequency)))
 
     if mp.am_master():
         if eps_parameters['contour']:
@@ -543,7 +543,7 @@ def plot_fields(sim, ax=None, fields=None, output_plane=None, field_parameters=N
 def plot2D(sim, ax=None, output_plane=None, fields=None, labels=False,
            eps_parameters=None, boundary_parameters=None,
            source_parameters=None, monitor_parameters=None,
-           field_parameters=None, resolution=None,
+           field_parameters=None, frequency=None, resolution=None,
            plot_eps_flag=True, plot_sources_flag=True,
            plot_monitors_flag=True, plot_boundaries_flag=True):
 
@@ -551,6 +551,23 @@ def plot2D(sim, ax=None, output_plane=None, fields=None, labels=False,
     if ax is None and mp.am_master():
         from matplotlib import pyplot as plt
         ax = plt.gca()
+
+    # Determine a frequency to plot all epsilon
+    if frequency is None:
+        try:
+            # Continuous sources
+            frequency = sim.sources[0].frequency
+        except:
+            try:
+                # Gaussian sources
+                frequency = sim.sources[0].src.frequency
+            except:
+                try:
+                    # Custom sources
+                    frequency = sim.sources[0].src.center_frequency
+                except:
+                    # No sources
+                    frequency = 0
 
     # validate the output plane to ensure proper 2D coordinates
     from meep.simulation import Volume
@@ -560,7 +577,8 @@ def plot2D(sim, ax=None, output_plane=None, fields=None, labels=False,
     # Plot geometry
     if plot_eps_flag:
         ax = plot_eps(sim, ax, output_plane=output_plane,
-                      eps_parameters=eps_parameters, resolution=resolution)
+                      eps_parameters=eps_parameters, resolution=resolution,
+                      frequency=frequency)
 
     # Plot boundaries
     if plot_boundaries_flag: