--- a/src/meepgeom.hpp
+++ b/src/meepgeom.hpp
@@ -281,7 +281,8 @@ void get_epsilon_grid(geometric_object_list gobj_list,
                       int nx, const double *x,
                       int ny, const double *y,
                       int nz, const double *z,
-                      double *grid_vals);
+                      std::complex<double> *grid_vals,
+                      double frequency = 0);
 void init_libctl(material_type default_mat, bool ensure_per,
                  meep::grid_volume *gv, vector3 cell_size, vector3 cell_center,
                  geometric_object_list *geom_list);