… doc/docs/Python_User_Interface.md
… src/meepgeom.cpp