Closes #1824.

Removed the `frequency` parameter from `plot2D` since this cannot be passed to `get_epsilon_grid`. As a result, this API change is not backwards compatible. As a workaround, we can leave the `frequency` parameter in but just mention in the documentation that it is simply ignored.

cc @joamatab