@oskooi 2021-11-13
For the record, there was a single test failure unrelated to this PR that went away after rerunning. The symmetry unit test of `test_material_grid.py` was failing a check in the sixth decimal digit for single precision:
```
======================================================================
FAIL: test_symmetry (__main__.TestMaterialGrid)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/home/runner/work/meep/meep/build/meep-1.22.0-beta/_build/sub/python/../../../python/tests/test_material_grid.py", line 161, in test_symmetry
    self.assertAlmostEqual(tran_nosym, tran_sym)
AssertionError: 20.27522531963953 != 20.275228328448193 within 7 places (3.0088086617752197e-06 difference)
```
It might therefore be useful to increase the tolerance slightly for this test to avoid this failure in the future.