@smartalecH 2021-11-13
Is it just as fast when plotting a 2D cross section of a 3D simulation?