@oskooi 2021-11-14
> Is it just as fast when plotting a 2D cross section of a 3D simulation?

Taking the example above and extending it to 3d, the difference in runtimes at a resolution of 320 for this PR compared to master is 0.5773 s vs. 87.5444 s. This translates to a speedup of a factor of more than 150. Using the revamped `plot2D` in this PR to generate a 2D cross section of a 3D cell seems to be just as fast as the 2D cell shown previously (both roughly ~0.5 s).

![plot2D_comparison_3d](https://user-images.githubusercontent.com/7152530/141664421-3b3e0664-de39-4dc0-ba47-8e007439e25a.png)

**PR 1827**
```
Using MPI version 3.1, 1 processes
     cylinder, center = (0.35,0.1,0)
          radius 0.1, height 0.5, axis (0, 0, 1)
          dielectric constant epsilon diagonal = (2.25,2.25,2.25)
     block, center = (-0.15,-0.2,0)
          size (0.2,0.24,0.8)
          axes (1,0,0), (0,1,0), (0,0,1)
          dielectric constant epsilon diagonal = (6.25,6.25,6.25)
     block, center = (-0.2,0.2,0)
          size (0.4,0.4,0.3)
          axes (1,0,0), (0,1,0), (0,0,1)
Geometric-object bounding-box tree:
     box (-0.5..0.5, -0.5..0.5, 0..0)
          bounding box (-0.4..0, 0..0.4, -0.15..0.15)
          shift object by (0, 0, 0)
          block, center = (-0.2,0.2,0)
               size (0.4,0.4,0.3)
               axes (1,0,0), (0,1,0), (0,0,1)
          bounding box (-0.25..-0.05, -0.32..-0.08, -0.4..0.4)
          shift object by (0, 0, 0)
          block, center = (-0.15,-0.2,0)
               size (0.2,0.24,0.8)
               axes (1,0,0), (0,1,0), (0,0,1)
          bounding box (0.25..0.45, 0..0.2, -0.25..0.25)
          shift object by (0, 0, 0)
          cylinder, center = (0.35,0.1,0)
               radius 0.1, height 0.5, axis (0, 0, 1)
Geometric object tree has depth 1 and 3 object nodes (vs. 3 actual objects)

Elapsed run time = 0.5773 s
```

**master**
```
Using MPI version 3.1, 1 processes
-----------
Initializing structure...
time for choose_chunkdivision = 3.6012e-05 s
Working in 3D dimensions.
Computational cell is 1 x 1 x 1 with resolution 320
     cylinder, center = (0.35,0.1,0)
          radius 0.1, height 0.5, axis (0, 0, 1)
          dielectric constant epsilon diagonal = (2.25,2.25,2.25)
     block, center = (-0.15,-0.2,0)
          size (0.2,0.24,0.8)
          axes (1,0,0), (0,1,0), (0,0,1)
          dielectric constant epsilon diagonal = (6.25,6.25,6.25)
     block, center = (-0.2,0.2,0)
          size (0.4,0.4,0.3)
          axes (1,0,0), (0,1,0), (0,0,1)
Geometric-object bounding-box tree:
     box (-0.503125..0.503125, -0.503125..0.503125, -0.503125..0.503125)
          bounding box (-0.4..0, 0..0.4, -0.15..0.15)
          shift object by (0, 0, 0)
          block, center = (-0.2,0.2,0)
               size (0.4,0.4,0.3)
               axes (1,0,0), (0,1,0), (0,0,1)
          bounding box (-0.25..-0.05, -0.32..-0.08, -0.4..0.4)
          shift object by (0, 0, 0)
          block, center = (-0.15,-0.2,0)
               size (0.2,0.24,0.8)
               axes (1,0,0), (0,1,0), (0,0,1)
          bounding box (0.25..0.45, 0..0.2, -0.25..0.25)
          shift object by (0, 0, 0)
          cylinder, center = (0.35,0.1,0)
               radius 0.1, height 0.5, axis (0, 0, 1)
Geometric object tree has depth 1 and 3 object nodes (vs. 3 actual objects)
subpixel-averaging is 17.152% done, 19.3209 s remaining
subpixel-averaging is 30.5805% done, 9.08038 s remaining
subpixel-averaging is 43.1867% done, 5.26365 s remaining
subpixel-averaging is 59.4996% done, 2.72321 s remaining
subpixel-averaging is 77.5109% done, 1.16075 s remaining
subpixel-averaging is 91.4617% done, 0.374134 s remaining
subpixel-averaging is 17.1149% done, 19.3729 s remaining
subpixel-averaging is 30.3951% done, 9.1621 s remaining
subpixel-averaging is 42.9373% done, 5.31743 s remaining
subpixel-averaging is 59.0481% done, 2.77456 s remaining
subpixel-averaging is 77.1773% done, 1.18323 s remaining
subpixel-averaging is 91.0708% done, 0.392198 s remaining
subpixel-averaging is 17.1217% done, 19.3657 s remaining
subpixel-averaging is 30.5299% done, 9.1024 s remaining
subpixel-averaging is 43.2002% done, 5.26053 s remaining
subpixel-averaging is 59.4356% done, 2.7301 s remaining
subpixel-averaging is 77.4469% done, 1.16553 s remaining
subpixel-averaging is 91.2022% done, 0.386007 s remaining
time for set_epsilon = 86.1741 s
-----------

Elapsed run time = 87.5444 s

```